q2h done

+++ q2_initialization.py
@@ -25,8 +25,8 @@ def xavier_weight_init():
             out: tf.Tensor of specified shape sampled from the Xavier distribution.
         """
         shape=list(shape)
-        epsilon = np.sqrt(6) / np.sum(shape)
-        out = tf.random_uniform(shape, minval=-epsilon, maxval=epsilon)
+        epsilon = np.sqrt(6.0) / np.sum(shape)
+        out = tf.random_uniform(shape, dtype=tf.float32, minval=-epsilon, maxval=epsilon)
         return out
     # Returns defined initializer function.
     return _xavier_initializer
